python-backend/requirements.txt
absl-py==2.3.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.30.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
deepface==0.0.93
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.72.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
libclang==18.1.1
lz4==4.4.4
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.3
numpy==2.1.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.55
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pgvector==0.4.1
pillow==11.2.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
retina-face==0.0.17
rich==14.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.19.0
tomli==2.2.1
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
